+++ packages/docs/pages/components/modal/_meta.ts
@@ -1,0 +1,8 @@
+export default {
+  'best-practices': 'Best Practices',
+  'modal-content': 'ModalContent',
+  'modal-dismiss': 'ModalDismiss',
+  'modal-footer': 'ModalFooter',
+  'modal-header': 'ModalHeader',
+  'modal-heading': 'ModalHeading',
+}