+++ packages/docs/pages/components/select/_meta.ts
@@ -1,0 +1,7 @@
+export default {
+  'select-item': 'SelectItem',
+  'select-item-check': 'SelectItemCheck',
+  'select-list': 'SelectList',
+  'select-popover': 'SelectPopover',
+  'select-trigger': 'SelectTrigger',
+}