+++ packages/docs/pages/guides/code/_meta.ts
@@ -1,0 +1,8 @@
+export default {
+  'development-guideline': 'Development Guideline',
+  'component-model': 'Component Model',
+  'storybook-guideline': 'Storybook Guideline',
+  'documentation-guideline': 'Documentation Guideline',
+  'test-guideline': 'Test Guideline',
+  'code-styleguide': 'Code Styleguide',
+}