+++ packages/docs/pages/guides/design/_meta.ts
@@ -1,0 +1,5 @@
+export default {
+  'handoff-requirements': 'Handoff Requirements',
+  component: 'Component',
+  icon: 'Icon',
+}