+++ packages/shoreline/src/components/date-picker/date-picker.tsx
@@ -22,18 +22,25 @@ import { useStore } from '@vtex/shoreline-utils'
  * <DatePicker />
  */
 export function DatePicker<T extends DateValue>(props: DatePickerProps<T>) {
-  const { className, id: defaultId, error: defaultError, ...domProps } = props
+  const {
+    className,
+    id: defaultId,
+    error: defaultError,
+    portal = true,
+    ...domProps
+  } = props
   const state = useDatePickerState(domProps)
   const ref = useRef(null)
   const anchorRef = useRef<HTMLDivElement>(null)
-  const { groupProps, labelProps, fieldProps, buttonProps, calendarProps } =
-    useDatePicker(domProps, state, ref)
 
   const store = useFieldContext()
   const { id: contextId, error: contextError } = useStore(store, (s) => s)
 
   const id = defaultId || contextId
   const error = defaultError || contextError
+
+  const { groupProps, labelProps, fieldProps, buttonProps, calendarProps } =
+    useDatePicker({ ...domProps, id }, state, ref)
 
   return (
     <PopoverProvider
@@ -60,6 +67,7 @@ export function DatePicker<T extends DateValue>(props: DatePickerProps<T>) {
                 label={buttonProps['aria-label']}
                 aria-describedby={buttonProps['aria-describedby']}
                 variant="tertiary"
+                disabled={buttonProps.isDisabled}
               >
                 <IconCalendarBlank />
               </IconButton>
@@ -68,6 +76,7 @@ export function DatePicker<T extends DateValue>(props: DatePickerProps<T>) {
         </div>
       </div>
       <Popover
+        portal={portal}
         getAnchorRect={() => {
           if (anchorRef?.current) {
             return anchorRef.current.getBoundingClientRect()
@@ -92,6 +101,11 @@ export interface DatePickerOptions<T extends DateValue>
    * Wether has error
    */
   error?: boolean
+  /**
+   * Should activate portal
+   * @default true
+   */
+  portal?: boolean
 }
 
 export type DatePickerProps<T extends DateValue> = DatePickerOptions<T>